requirements.txt
google-auth>=2.0
requests
